Rename 'Component' to 'Segment'
