+++ src/main/kotlin/feature/segment/StateSegment.kt
@@ -1,12 +1,12 @@
-package feature.component
+package feature.segment
 
-import feature.FileType
 import feature.CreateFileResult
+import feature.FileType
 import feature.createEiffelFile
 import util.createOrGetSubdirectory
 
-val stateComponent: Component = { templateManager, properties, _, featureName ->
-    createEiffelComponent { skippedFiles ->
+val stateSegment: Segment = { templateManager, properties, _, featureName ->
+    createEiffelSegment { skippedFiles ->
         createOrGetSubdirectory("state").run {
             createEiffelFile(FileType.Action(featureName), templateManager, properties).let {
                 if (it is CreateFileResult.AlreadyExists) skippedFiles.add(it.name)
