+++ src/main/kotlin/feature/segment/UISegment.kt
@@ -1,12 +1,12 @@
-package feature.component
+package feature.segment
 
-import feature.FileType
 import feature.CreateFileResult
+import feature.FileType
 import feature.createEiffelFile
 import util.createOrGetSubdirectory
 
-val uiComponent: Component = { templateManager, properties, rootDirectory, featureName ->
-    createEiffelComponent { skippedFiles ->
+val uiSegment: Segment = { templateManager, properties, rootDirectory, featureName ->
+    createEiffelSegment { skippedFiles ->
         createOrGetSubdirectory("ui").run {
             createEiffelFile(FileType.Fragment(featureName), templateManager, properties).let {
                 if (it is CreateFileResult.AlreadyExists) skippedFiles.add(it.name)
